pyproject.toml
[tool.poetry]
name = "conda-forge-warning"
version = "0.1.0"  # semantic-release
description = "Conda-Forge Warning aims to present conda-forge package without maintenance"
authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>"]
packages = [
  {include = "cf_warning", from="src"},
]
license = "BSD 3 Clause"
exclude = [
  ".git/*",
  ".env*",
]
include = ["src/cf_warning/py.typed"]


[tool.poetry.dependencies]
python = "^3.8.1"
pandas = "^2.0.3"
gql = "^3.4.1"
aiohttp = "^3.8.5"
atpublic = "^4.0"
python-dotenv = ">=0.21"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
urllib3 = "<2"  # fix poetry issues
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
coverage = "^7.2.7"
black = "^23.3.0"
pre-commit = "^3.3.2"
ruff = "^0.0.278"
mypy = "^1.3.0"
bandit = "^1.7.5"
pydocstyle = "^6.3.0"
vulture = "^2.7"
mccabe = "^0.6.1"
ipython = "<8"
ipykernel = ">=6.0.0"
Jinja2 = ">=1.8.0"
mkdocs = "^1.4.3"
mkdocs-exclude = "^1.0.2"
mkdocs-jupyter = "^0.24.1"
mkdocs-literate-nav = "^0.6.0"
mkdocs-macros-plugin = ">=0.7.0,<1"
mkdocs-material = "^9.1.15"
mkdocstrings = "^0.21.2"
mkdocstrings-python = "^1.1.2"
makim = "^1.8.2"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.black]
line-length = 79
target-version = ["py38"]
force-exclude = '''(?x)(
    docs/*
  | .*\\.egg-info
)'''  # TOML's single-quoted strings do not require escaping backslashes

[tool.bandit]
exclude_dirs = ["tests"]
targets = "./"

[tool.vulture]
exclude = ["tests"]
ignore_decorators = []
ignore_names = []
make_whitelist = true
min_confidence = 80
paths = ["./"]
sort_by_size = true
verbose = false

[tool.ruff]
line-length = 79
force-exclude = true
src = ["./"]
exclude = [
  'docs',
]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "D",   # pydocstyle
  "YTT", # flake8-2020
  "RUF", # Ruff-specific rules
  "I001",  # isort
]
fixable = ["I001"]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.mypy]
no_strict_optional = false
